LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/prefetch-single.https_from_protocol=http&to_protocol=https-expected.txt

CONSOLE MESSAGE: Prefetch request denied: not same origin as document
CONSOLE MESSAGE: Prefetch request denied: not same origin as document

FAIL test single https url prefetch from a http url assert_in_array: Prefetch should work for HTTPS urls. value undefined not in array ["prefetch", "prefetch;anonymous-client-ip"]
